requirements.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
joblib==1.4.2
MetaTrader5==5.0.4993
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
ta==0.11.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
